new_requirements.txt
fastapi
pymysql
mysql-connector
tqdm
pandas
openpyxl
mysql-connector-python
uvicorn
python-multipart
aiofiles
jinja2
requests_toolbelt